Add files via upload
This internship report provides a comprehensive overview of my internship experience at the AK INFOPARK on Mobile Application Development in Nagercoil, Tamil Nadu. The internship provided me with an invaluable opportunity to gain practical knowledge and hands-on experience in the world of code, design and user experience operations, contributing to my professional understanding and growth.